requirements.txt

psycopg2-binary
pydantic
pyyaml
sqlalchemy == 1.4
fastapi
uvicorn 

